+++ packages/exchange-consumer/src/index.ts
@@ -1,6 +1,7 @@
 export * from './consumers'
 export * from './decomposers'
 export * from './composers'
+export * from './proto-composers'
 export * from './streams'
 export * from './transformers'
 export * from './types'