+++ packages/spot-consumer/src/index.ts
@@ -1,5 +1,6 @@
 export * from './consumers'
 export * from './transformers'
 export * from './composers'
+export * from './proto-composers'
 export * from './streams'
 export * from './types'